requirements.txt
pandas
matplotlib
Pillow
python-dotenv

ipython
google-genai
tabulate

requests
qrcode[pil]

fastapi
uvicorn
sqlalchemy
jinja2
markdown
